implementado os testes unitarios e integrado com o travis
