+++ src/main/java/org/cesed/map/meuchurrascoapp/base/BaseBean.java
@@ -19,4 +19,5 @@ public abstract class BaseBean implements Serializable{
 	public boolean equals(Object obj){
 		return EqualsBuilder.reflectionEquals(this, obj);
 	}
+	
 }
