+++ src/main/java/org/cesed/map/meuchurrascoapp/notificacao/EmailNotification.java
@@ -15,7 +15,7 @@ import com.sendgrid.SendGrid;
 public class EmailNotification {
 	
 	public void sendEmail(String toEmail, Evento evento) throws IOException{
-		Email from = new Email("davi.leal737@hotmail.com");
+		Email from = new Email("davi.leal737@gmail.com");
 	    String subject = "Meu churrasco App - Convite";
 	    Email to = new Email(toEmail);
